requirements.txt
python-dotenv==1.1.0
fireworks-ai==0.16.1
pytest==8.3.5
pre_commit==4.2.0
pydantic==2.11.4